Refactor displayed category of monos (#1906)
This follows the second suggestion of #1904 . The displayed category of
monomorphisms is defined as a displayed category over the category of
morphisms. The only disadvantage to this change is that one cannot
directly switch between morphisms in `C / x` and `C /m x`. The advantage
is that the definition of the inclusion of `C /m x` into `C / x` is
automatic, and that the approach is more principled.