Add basic definitions and results regarding pre-bilattices.
Pre-bilattices are structures comprising two lattices (over the same set).

First, this commit defines pre-bilattices, bounded prebilattices,
and the displayed category of pre-bilattices.

Moreover, it defines two special cases:
interlaced and distributive pre-bilattices.

Also, it defines the construction of product pre-bilattices
and proves that product pre-bilattices are interlaced
and that every interlaced pre-bilattice is isomoprhic to a product one
(representation theorem).

The pre-bilattice FOUR is included as an example of a pre-bilattice.