Refactoring development of comprehension categories and internal language theorems (#1948)
This PR contains several refactors and one addition to the development
on comprehension categories. The content of this PR is as follows.
- Instead of require left adjoints to exist for every substitution,
sigma-types now only require them to exist for weakenings.
- The same for pi-types. Some proofs became more complicated because of
this change.
- Local properties now refer to finitely complete categories instead. As
a consequence, subobject classifiers can now be represented as a local
property. This allows saving some code, because the biequivalence for
subobject classifiers is now an instance of the one of local properties.
- Several constructions for displayed pseudofunctors/transformations
required too many assumptions (in the file DispBuilders), and this has
been made more efficients.
- I added parameterized natural numbers object, showed that they give a
local property, and I added natural numbers to the internal language
theorem

Note: the changes to sigma-types and pi-types also saved some code.
Before the displayed bicategories of sigma-types and of pi-types were
defined on top of the bicategory of categories+fibrations+terminal
object. Now they are defined directly on (full) comprehension
categories. This means that less liftings are required.

---------

Co-authored-by: Benedikt Ahrens <benedikt.ahrens@gmail.com>