More on the setgroupoid model (#1963)
In my previous PR, I defined a univalent (non-full) comprehension
category of setgroupoids. In that comprehension category, types in a
context `G` are given by pseudofunctors from `G` to the bicategory of
setgroupoids. In this PR, I define a variation where types in a context
`G` are given by functors from `G` to the category of setgroupoids. This
is done by adding suitable splitness assumptions.