Autogenerate docs (#2032)
I believe it would be good to automatically generate rocqdoc (and other
kinds of derived content) from the code at `master`, and combine that
with the `*.md` files to get a website, very much like the way it works
for [the HoTT
library](https://github.com/HoTT/Coq-HoTT/blob/master/.github/workflows/ci.yml).

The last step of this PR fails because of permission issues (actions
triggered by a commit in my fork cannot have write access on the UniMath
repository). However, this afternoon I triggered the same action from
the newly created
[UniMath/autogenerate-docs](https://github.com/UniMath/UniMath/tree/autogenerate-docs)
branch, which correctly gathered everything we need (apart from the
jekyll config) on the
[gh-pages](https://github.com/UniMath/UniMath/tree/gh-pages) branch.

The current version of my full proposal (copied from the approach of the
HoTT repository) is as follows:
1. People write code for UniMath, and commit it via pull requests. At
some point, this results in a commit on the master branch.
2. Currently, this triggers a CI action to compile the code. My idea is
to have this CI action also compile rocqdoc(, alectryon, a dependency
graph, etc)
3. After the documentation has been compiled, the CI action pushes this
via a single commit to the `gh-pages` branch. We can either configure it
to save earlier versions in the git history, or to overwrite the commits
each time and keep the history "clean".
4. Currently, [the UniMath site](https://unimath.github.io) is published
from the `master` branch. My proposal is to publish it from the
`gh-pages` branch. This means that every time a commit to `gh-pages` is
made, the site will update with the new docs.

Do you agree that this would be a good thing to have, and that this is a
good implementation?