Remove CONTENTS.md (#2036)
This PR is filed as a suggestion. It removes the file `CONTENTS.md` as I
think it is replaced by the [rocqdoc
TOC](https://unimath.github.io/UniMath/gen/rocqdoc/toc). It also removes
the sanity check related to this file, but leaves the code for
generating `CONTENTS.md` in place for now. Lastly, it adds a line or two
to the README, linking to the rocqdoc (toc) and alectryon docs.

Closes #862, closes #1634 by making them redundant.