Rezk Completion For Topoi (#2052)
This PR contains material about the Rezk completion.
More precisely, it is proven that the universal arrows/left adjoint
given by the Rezk completion lifts to
(a) pretopoi
(b) elementary topoi
(c) finite colimits

The material for result (b) was already in UniMath, with the exception
of the definition of the bicategory of elementary topoi.
Result (c) was also already in UniMath, with the exceptions of the
statement "if F : C -> D a weak equivalence, and C is closed under
colimits of a diagram, then D is also closed under that diagram" and the
bicategorical statement.
Results (a) and (b) have been extended to include natural numbers
objects.
To prove result (a), the Rezk completion for extensive categories
remained to be done.

Edit: The material about extensiveness and pretopoi has now been
removed.

---------

Co-authored-by: kfwullaert <kwullaer@hotmail.com>
Co-authored-by: Niels van der Weide <nmvdw@proton.me>
Co-authored-by: Niels van der Weide <nnmvdw@gmail.com>