Universes in categories with finite limits and comprehension categories (#2048)
This PR presents the following (it is large and it went a bit out of
hand, sorry for that):
- A construction of the univalent bicategory of categories with finite
limits and universes
- A construction of the univalent bicategory of comprehension categories
categories with universes

Along the way I develop more infrastructure for comprehension
categories, and, in particular, the coercion operation, which allows us
to view morphisms between types as a subtyping relation. I also
characterize universes in categories with finite limits to morphisms,
which connects it to the notion of universe as developed, for instance,
by Joyal&Moerdijk.